+++ src/main/java/localgaji/albbaim/auth/AuthService.java
@@ -6,7 +6,7 @@ import localgaji.albbaim.auth.user.UserService;
 import localgaji.albbaim.auth.oauth.kakaoAuth.KakaoAuth;
 import localgaji.albbaim.auth.oauth.kakaoAuth.KakaoAuthService;
 import localgaji.albbaim.__core__.auth.TokenProvider;
-import localgaji.albbaim.auth.user.userDTO.RequestAuth;
+import localgaji.albbaim.auth.authDTO.RequestAuth;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
